backend/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1

# Testing
pytest==8.3.3
pytest-cov==6.0.0
pytest-asyncio==0.24.0
httpx==0.27.2

# Development
python-dotenv==1.0.1
